+++ packages/position/src/react/__specs__/index.spec.js
@@ -1,0 +1,3 @@
+describe('Position', () => {
+  it('renders', () => expect(true).toBeTruthy())
+})